requirements.txt
fastapi
boto3
uvicorn
torch
pillow
python-multipart
facenet_pytorch